Running with prophecy refiner -- not tested yet Note: This code is super messy -- especially with the use of const Things are changing in all kinds of places. It's not really clear where the right place to store things are.
e.g. it would be nice if the abstractor retained all the information from abstracting and
then we could just query it without modifying it more, but we have ugly implementation details
like "orig_sorts" that needs to be modified as we add new indices (from prophecy refinement) and
that's currently in the abstractor still.