Make violated_axioms a set
There are some cases where an axiom can be enumerated twice (in the future,
we should avoid that). I think this happens when an equality appears in both
init and trans.

For some reason, we were getting a failure in the model evaluation, when
we asserted an axiom twice and then later tried to get the value.

I'm not sure why that would be the case, so unfortunately I'm not positive
this *really* fixed the issue...